requirements/base.txt
argcomplete==1.10.0
beautifulsoup4==4.8.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.2
defusedxml==0.6.0
Django==2.2.5
django-crispy-forms==1.7.2
django-environ==0.4.5
django-filter==2.2.0
django-flatpickr==1.0.1
djangorestframework==3.10.3
djangorestframework-datatables-editor==0.3.2
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
html5lib==1.0.1
idna==2.8
IMAPClient==2.1.0
lxml==4.4.1
Markdown==3.1.1
olefile==0.46
pdfminer.six==20181108
Pillow==6.1.0
pycparser==2.19
pycryptodome==3.9.4
Pyphen==0.9.5
python-dateutil==2.8.0
python-pptx==0.6.18
pytz==2019.2
requests==2.22.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
sqlparse==0.3.0
textract==1.6.3
tinycss2==1.0.2
tzlocal==1.5.1
urllib3==1.25.7
WeasyPrint==50
webencodings==0.5.1
xlrd==1.2.0
XlsxWriter==1.2.6
